fix: update upload outlet_image, outlet_facility

+++ app/Repository/OutletFasilityRepository.php
@@ -2,15 +2,20 @@
 namespace App\Repository;
 
 use App\Interface\OutletFasilityInterface;
+use App\Models\Outlet;
 use App\Models\OutletFacility;
 use Illuminate\Contracts\Cache\Store;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Storage;
 
 class OutletFasilityRepository implements OutletFasilityInterface {
     public function index()
     {
-        return view('outlet-fasility');
+        $data = Outlet::all();
+        return view('outlet-fasility', [
+            'data' => $data
+        ]);
     }
 
     public function store(Request $request)
@@ -29,7 +34,7 @@ class OutletFasilityRepository implements OutletFasilityInterface {
 
     public function show()
     {
-        $outletFasility = OutletFacility::all();
+        $outletFasility = OutletFacility::with('outlet')->get();
 
         return response()->json([
             'data' => $outletFasility
@@ -45,16 +50,39 @@ class OutletFasilityRepository implements OutletFasilityInterface {
     public function update($id, Request $request)
     {
         $outletFasility = OutletFacility::findOrFail($id);
-        $outletFasility->update($request->all());
+        if(isset($request->icon_outlet_facility)){
+            File::delete(public_path($outletFasility->icon_outlet_facility));
 
-        return response()->json(['status' => true]);
+            $storeImage = Storage::put('public/icon', $request->file('icon_outlet_facility'));
+            $storeImageUrl = Storage::url($storeImage, 'public/icon');
+
+            $outletFasility->update([
+                'id_outlet' => $request->id_outlet, 
+                'icon_outlet_facility' => $storeImageUrl,
+                'description_outlet_facility' => $request->description_outlet_facility
+            ]);
+
+        }else{
+            $outletFasility->update([
+                'id_outlet' => $request->id_outlet,
+                'description_outlet_facility' => $request->description_outlet_facility
+            ]);
+        }
+        return response()->json([
+            'status' => true
+        ]);
     }
 
     public function destroy($id)
     {
         $outletFasility = OutletFacility::findOrFail($id);
-        $outletFasility->delete();
+        if(File::exists(public_path($outletFasility->icon_outlet_facility))){
+            File::delete(public_path($outletFasility->icon_outlet_facility));
+            $outletFasility->delete();
+            return response()->json(['status' => true]);
+        }else{
+            return response()->json(['status' => false]);
+        }
 
-        return response()->json(['status' => true]);
     }
 }
